refactored keyboard navigation
